feat: Add products handler and tests

+++ migrations/sqls/20220418112932-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR NOT NULL,
-    price INTEGER NOT NULL,
+    price NUMERIC(12,2) NOT NULL,
     category VARCHAR
 );
